Tighten up stream integration tests
Reject messages which contain errors when running the integration
test suite.

This comes in handy when one wants to attach some processors to
the input(s) in integration tests for performing additional
validations such as throwing an error if some expected metadata
field isn't present or has the wrong value.

Signed-off-by: Mihai Todor <todormihai@gmail.com>