Add `error_source_*` bloblang functions
- `error_source_name()`: Returns the name of the component which
raised the error.
- `error_source_label()`: Returns the label of the component
which raised the error.
- `error_source_path()`: Returns the config path of the
conponent which raised the error.

One use case for these is to determine which processor generated
an error within a `try` processor.

I thought about having only `error_source_label()` and returning
the path when the label is not set, but it feels handy to be able
to quickly print both the label and the path.

When using the same processor resource multiple times in a config,
these functions won't help users figure out which config section
caused the error, since the path is relative to
`processor_resources`. Templates don't have this issue because the
path will be derived from the config itself and one can now even
namespace the subcomponent label using the new `@label` metadata
in the template `mapping`.

Signed-off-by: Mihai Todor <todormihai@gmail.com>