Fix http_client input connect bug
This issue was uncovered because the `http_client` components
would not try to reconnect if the internal codec couldn't be
initialised. The fix addresses the issue at the scanner level.

`h.codecCtor.Create` will return a `*service.OwnedScanner`, but
`h.codec` has type `codec.DeprecatedFallbackStream`, so if we get
an error from `h.codecCtor.Create`, the `h.codec != nil` check in
`Connect` will be true, so Benthos will think the input is
connected. This can happen if the codec specification isn't
compatible with the received data.

Signed-off-by: Mihai Todor <todormihai@gmail.com>